Refactor DailyMedication into a new package
