+++ src/test/java/meditracker/DailyMedicationManagerTest.java
@@ -1,12 +1,14 @@
 package meditracker;
 
 import meditracker.command.AddCommand;
+import meditracker.dailyMeds.DailyMedication;
+import meditracker.dailyMeds.DailyMedicationManager;
 import meditracker.exception.ArgumentNotFoundException;
 import meditracker.exception.DuplicateArgumentFoundException;
 import meditracker.exception.FileReadWriteException;
 import meditracker.medication.Medication;
 import meditracker.medication.MedicationManager;
-import meditracker.medication.SubDailyManager;
+import meditracker.dailyMeds.SubDailyManager;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
