+++ src/test/java/meditracker/command/TakeCommandTest.java
@@ -1,12 +1,12 @@
 package meditracker.command;
 
-import meditracker.DailyMedication;
-import meditracker.DailyMedicationManager;
+import meditracker.dailyMeds.DailyMedication;
+import meditracker.dailyMeds.DailyMedicationManager;
 import meditracker.exception.ArgumentNotFoundException;
 import meditracker.exception.FileReadWriteException;
 import meditracker.exception.DuplicateArgumentFoundException;
 import meditracker.medication.MedicationManager;
-import meditracker.medication.SubDailyManager;
+import meditracker.dailyMeds.SubDailyManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
